requirements.txt
flask>=2.0.0
werkzeug>=2.0.0
biopython==1.85
pandas==2.0.3
numpy==1.24.3
plotly>=5.18.0
scikit-learn>=1.0.0
seaborn>=0.12.0
matplotlib>=3.7.0
dash>=2.14.0
requests==2.26.0
